lib: move lvgl fonts and themes to lvgl addons lib
