+++ lib/register_bsb_common_libs.scons
@@ -8,8 +8,7 @@ for libname in [
     "canvas_oled",
     "lwip",
     "lvgl",
-    "lvgl_fonts",
-    "lvgl_themes",
+    "lvgl_addons",
     "network",
     "mongoose",
     "fw_updater_opts",
