+++ targets/f20/target.json
@@ -103,8 +103,7 @@
         "fatfs",
         "lwip",
         "lvgl",
-        "lvgl_fonts",
-        "lvgl_themes",
+        "lvgl_addons",
         "network",
         "mongoose",
         "canvas_oled",
